pyproject.toml
[tool.poetry]
name = "Run-TOGT-Planner"
version = "1.0.0"
description = "This Python package provides an interface to run the Time-Optimal Gate-Traversing (TOGT) Planner."
authors = ["YamakenElf <810658920qq@gmail.com>"]
license = "MIT"
readme = "pypi_description.md"
repository = "https://github.com/KafuuChikai/Run-TOGT-Planner"
packages = [
    { include = "run_togt_planner" }
]

[tool.poetry.dependencies]
python = ">=3.8"
numpy = ">=1.19.0"
pyyaml = ">=5.4.1"
"ruamel.yaml" = ">=0.15.100"
matplotlib = ">=3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"